Backend/requirements.txt
aiohttp==3.8.6
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.1
cryptography==41.0.4
defusedxml==0.7.1
distlib==0.3.7
Django==4.1.4
django-allauth==0.56.1
django-cors-headers==4.3.0
django-phonenumber-field==7.3.0
django-rest-auth==0.9.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
filelock==3.12.4
frozenlist==1.4.0
gunicorn==21.2.0
idna==3.4
multidict==6.0.4
oauthlib==3.2.2
packaging==23.2
phonenumbers==8.13.31
Pillow==10.1.0
platformdirs==3.11.0
psycopg2==2.9.9
pycparser==2.21
PyJWT==2.8.0
python-decouple==3.8
python3-openid==3.2.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.4
twilio==8.10.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
virtualenv==20.24.5
whitenoise==6.6.0
yarl==1.9.2
